Remove redundant @CorePlatformApi from Inet6Address.
This is unnecessary since the class is part of the public API (see
https://developer.android.com/reference/java/net/Inet6Address) because
it has no @hide. The same goes for the getScopeId() method. The effect
of this annotation is to cause the class's public methods to appear in
the core_platform_api current.txt, which is misleading.

(There are two fields in this class which are marked @hide and
@CorePlatformApi. This is presumably intentional, and this change does
not affect them, they remain part of the core platform API.)

Test: treehugger
Bug: 154796679
Change-Id: I08dc650e6e33b138005676070888d6b33cb574d6